requirements.txt
boto3==1.26.151
botocore==1.29.151
jmespath==1.0.1
Pillow==9.5.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
